# speed up deletes and add onDeletesFlushed()