For issue https://github.com/ssb-ngi-pointer/ssb-db2/issues/334

@arj03 I don't think this PR is ready but I want to ask you a couple questions.

I'm using a batching strategy similar to `scheduleWrite` for the deletes. The results are pretty nice:

**Before:**

```
append 1000 records: 125.351ms
delete 1000 records: 897.869ms
```

**After:**

```
append 1000 records: 107.628ms
delete 1000 records: 19.421ms
```

I suspect it got faster than append because deletes use `buffer.fill` while appends use `buffer.copy`.

Issues:

- I added the new API `onFlushDelete` which is like `onDrain` but for deletes. I don't think this is great, I would have wanted that `onDrain` also waits for deletes to end, but it's tricky because `onDrain` just adds CBs to the `latestBlockIndex`, but deletes can happen on any block. Any thoughts about this?
- There are probably N corner cases with concurrency problems between append and delete. Like deletes on a "blockToBeWritten" is going to be postponed with `onDrain`, but not if the delete is for an older block while there are appends queued up for the lastBlockIndex. This is probably a mine field, have to walk carefully here, which is why the PR is a draft.