@arj03 2022-04-14
I do like them to be separate. It makes it easier to describe what they do and to reason about them. As for case no 2, I think it would be ok if deleting something in the latest non-flushed block will simply throw an error. In practice it would be a strange case to delete something that has not been written because indexes has to run on data that is written, otherwise in the case of a crash the indexes don't match the log.