@staltz 2022-04-14
Thanks for the review! I added commit https://github.com/ssb-ngi-pointer/async-append-only-log/pull/68/commits/c68eef36f25b76f044ab4c7323b5698196a3cae4