@staltz 2022-04-14
Question: can there be two concurrent `raf.read` for the same block? Say we call two `del` on the same block, and the block is not cached so it triggers two concurrent `getBlock` thus two `raf.read` which produces two different blockBuf.

When the 2nd `gotBlockForDelete` triggers, it would overwrite the blockBuf on `blocksWithDeletables`.