@arj03 2022-04-14
Good catch in latest commit. Merge away

Sendt fra ProtonMail-mobil

-------- Originalbesked --------
Til 14. apr. 2022 15.56, André Staltz skrev:

> Okay, I just added another commit that addresses the problem I just spoke of.
>
> —
> Reply to this email directly, [view it on GitHub](https://github.com/ssb-ngi-pointer/async-append-only-log/pull/68#issuecomment-1099214192), or [unsubscribe](https://github.com/notifications/unsubscribe-auth/AACVTWWADXRHD2WNF6FKQLDVFAP2NANCNFSM5TM64W3A).
> You are receiving this because you were mentioned.Message ID: ***@***.***>